requirements.txt
aiogram==3.14.0
betterlogging==1.0.0
orjson==3.10.11
python-dotenv==1.0.1
sqlalchemy
asyncpg
python-dotenv
requests
notion-client
aiosqlite
logging